Analysis of mutual information data -- R code buggy
The R code in "Lab notebook.Rmd" works only partially.  I can get the current analysis to work if I run it from the console, but not via knitting the notebook.  Bug shows up as different numbers of rows in a data.frame and in a vector that I'm trying to add to the data.frame as a new column.